+++ R/plot_effect.R
@@ -1,18 +1,48 @@
 #' @rdname plot_ecxsys
 #' @export
 plot_effect <- function(model,
-                        show_LL5_model = FALSE,
+                        which = NULL,
                         show_legend = FALSE,
                         xlab = "concentration",
-                        ylab = "effect") {
+                        ylab = "effect",
+                        main = NULL) {
     stopifnot(inherits(model, "ecxsys"))
-    temp <- adjust_smooth_concentrations(
-        model$curves,
-        model$conc_adjust_factor
-    )
+
+    curve_names <- names(model$curves)
+    valid_names <- curve_names[startsWith(curve_names, "effect")]
+    if (is.null(which)) {
+        which <- c("effect_tox_sys")
+        if (model$with_env) {
+            which <- c(which, "effect_tox_env_sys")
+        }
+    } else if ("all" %in% which) {
+        if (length(which) == 1) {
+            which <- valid_names
+        } else {
+            stop("'all' must not be combined with other curve names.")
+        }
+    } else if (!all(which %in% valid_names)) {
+        warning("Argument 'which' contains invalid names.")
+        if (!model$with_env && any(grepl("env", which, fixed = TRUE))) {
+            warning("'which' contains names with 'env' but the model was",
+                    " built without environmental effects.")
+        }
+        which <- which[which %in% valid_names]
+    }
+
+    temp <- adjust_smooth_concentrations(model)
     curves <- temp$curves
     log_ticks <- get_log_ticks(curves$concentration)
     concentration <- c(curves$concentration[1], model$args$concentration[-1])
+
+    legend_df <- data.frame(
+        text = character(),
+        pch = numeric(),
+        lty = numeric(),
+        col = character(),
+        order = numeric(),  # controls sorting of legend elements
+        stringsAsFactors = FALSE
+    )
 
     plot(
         NA,
@@ -22,93 +52,103 @@ plot_effect <- function(model,
         log = "x",
         xlab = xlab,
         ylab = ylab,
+        main = main,
         xaxt = "n",
-        bty = "L",
-        las = 1
+        yaxt = "n",
+        bty = "L"
     )
 
-    lines(
-        curves$concentration,
-        curves$effect_tox_sys,
-        col = "blue"
-    )
-    lines(
-        curves$concentration,
-        curves$effect_tox,
-        col = "deepskyblue",
-        lty = 2
-    )
     points(
         concentration,
         model$args$effect_tox_observed,
         pch = 16,
         col = "blue"
     )
-
+    legend_df[nrow(legend_df) + 1, ] <- list("tox (observed)", 16, 0, "blue", 1)
     if (model$with_env) {
-        lines(
-            curves$concentration,
-            curves$effect_tox_env_sys,
-            col = "red"
-        )
-        lines(
-            curves$concentration,
-            curves$effect_tox_env,
-            col = "orange",
-            lty = 2
-        )
         points(
             concentration,
             model$args$effect_tox_env_observed,
             pch = 16,
             col = "red"
         )
+        legend_df[nrow(legend_df) + 1, ] <- list("tox + env (observed)", 16, 0, "red", 2)
     }
-
-    if (show_LL5_model) {
+    # The lines are drawn in this order to ensure that dotted and dashed lines
+    # are on top of solid lines for better visibility.
+    if ("effect_tox_sys" %in% which) {
+        lines(
+            curves$concentration,
+            curves$effect_tox_sys,
+            col = "blue"
+        )
+        legend_df[nrow(legend_df) + 1, ] <- list("tox + sys", NA, 1, "blue", 5)
+    }
+    if ("effect_tox" %in% which) {
+        lines(
+            curves$concentration,
+            curves$effect_tox,
+            col = "deepskyblue",
+            lty = 2
+        )
+        legend_df[nrow(legend_df) + 1, ] <- list("tox", NA, 2, "deepskyblue", 4)
+    }
+    if ("effect_tox_LL5" %in% which) {
         lines(
             curves$concentration,
             curves$effect_tox_LL5,
             col = "darkblue",
-            lty = 4
+            lty = 3
         )
-        if (model$with_env) {
+        legend_df[nrow(legend_df) + 1, ] <- list("tox (LL5)", NA, 3, "darkblue", 3)
+    }
+    if (model$with_env) {
+        if ("effect_tox_env_sys" %in% which) {
+            lines(
+                curves$concentration,
+                curves$effect_tox_env_sys,
+                col = "red"
+            )
+            legend_df[nrow(legend_df) + 1, ] <- list("tox + env + sys", NA, 1, "red", 8)
+        }
+        if ("effect_tox_env" %in% which) {
+            lines(
+                curves$concentration,
+                curves$effect_tox_env,
+                col = "orange",
+                lty = 2
+            )
+            legend_df[nrow(legend_df) + 1, ] <- list("tox + env", NA, 2, "orange", 7)
+        }
+        if ("effect_tox_env_LL5" %in% which) {
             lines(
                 curves$concentration,
                 curves$effect_tox_env_LL5,
                 col = "darkred",
-                lty = 4
+                lty = 3
             )
+            legend_df[nrow(legend_df) + 1, ] <- list("tox + env (LL5)", NA, 3, "darkred", 6)
         }
     }
 
-    axis(1, at = log_ticks$major, labels = log_ticks$major_labels)
-    axis(1, at = log_ticks$minor, labels = FALSE, tcl = -0.25)
+    # The setting of col = NA and col.ticks = par("fg") is to prevent ugly line
+    # thickness issues when plotting as a png with type = "cairo" and at a low
+    # resolution.
+    axis(1, at = log_ticks$major, labels = log_ticks$major_labels,
+         col = NA, col.ticks = par("fg"))
+    axis(1, at = log_ticks$minor, labels = FALSE, tcl = -0.25,
+         col = NA, col.ticks = par("fg"))
     plotrix::axis.break(1, breakpos = temp$axis_break_conc)
+    axis(2, col = NA, col.ticks = par("fg"), las = 1)
 
     if (show_legend) {
-        legend_text <- c("tox", "tox + sys")
-        legend_pch <- c(NA, 16)
-        legend_lty <- c(2, 1)
-        legend_col <- c("deepskyblue", "blue")
-        if (model$with_env) {
-            legend_text <- c(legend_text, "tox + env", "tox + env + sys")
-            legend_pch <- c(legend_pch, NA, 16)
-            legend_lty <- c(legend_lty, 2, 1)
-            legend_col <- c(legend_col, "orange", "red")
-        }
-        if (show_LL5_model) {
-            legend_text <- c(legend_text, "tox (LL5)", "tox + env (LL5)")
-            legend_pch <- c(legend_pch, NA, NA)
-            legend_lty <- c(legend_lty, 4, 4)
-            legend_col <- c(legend_col, "darkblue", "darkred")
-        }
+        legend_df <- legend_df[order(legend_df$order), ]
         legend(
             "topright",
-            legend = legend_text,
-            pch = legend_pch,
-            lty = legend_lty,
-            col = legend_col
+            legend = legend_df$text,
+            pch = legend_df$pch,
+            lty = legend_df$lty,
+            col = legend_df$col
         )
     }
 }